refactor(ucm): relocate validate_tags to ucm/config/validate for rubric conformance
validate_tags was a validator living under ucm/config/mutator/ alongside
non-validator mutators. Move it to ucm/config/validate/ where the other
validators already live (reference_closure, naming, required_fields,
unique_resource_keys), for package-placement consistency with the rubric
and with bundle/config/mutator/validate/.

The mutator package's pre-move validate_tags_test.go carried the shared
loadUcm/summaries helpers used by the remaining mutator tests; extract
them into a new ucm/config/mutator/helpers_test.go so the package still
builds after the move. The validate package already has its own copy of
the same helpers in required_fields_test.go.

Call sites in ucm/phases/validate.go and ucm/config/validate/all.go
updated to import from the new location.
